Move construct when storing current value
For both the move assignment operator and emplace, the proposal P0323R10 states
to "move construct" the local variable from the object's current value and that
doing so "can't throw". However, for the copy assignment operator it states to
"construct" the local variable and that doing so "can throw". The copy
assignment operator is consistent with the move assignment operator and emplace
when move constructing the object's current value back from the local variable.

I cannot see a reason why move construction should not also be used to create
the local variable in the copy assignment operator. It could be a copy and
paste error in the proposal. Further, Sy Brand's implementation uses move
construction (see readme for link).