Make test class trivially default constructible
This test class is trivially destructible and is used to check that expected
is trivially destructible when it should be. Later this class will be used to
check that other special member functions of expected are trivial.

Making the default constructor trivial means that the state member cannot be
set. Related tests can rely on the default constructed value to assert that the
default constructor is called.